Removed gitHelpers and updated gitTasks
